fix #1195 & #1188, R: move some methods from ArgumentConverter to Binder
